+++ src/api/oraculPlaces.jsx
@@ -1,7 +1,7 @@
-const getOraculPlaces = async (accessToken) => {
+const fetchAllOraculPlaces = async (accessToken) => {
   try {
     const host = "http://localhost:5000";
-    const responseIncludeFields = "id,uuid,name,placeable_id,placeable_type";
+    const responseIncludeFields = "id,name,placeable_id,placeable_type";
     const response = await fetch(`${host}/api/v1/oracul_places?response_include_fields=${responseIncludeFields}&api_access_token=${accessToken}`, {
       method: "GET"
     });
@@ -12,4 +12,4 @@ const getOraculPlaces = async (accessToken) => {
   }
 };
 
-export { getOraculPlaces };
+export { fetchAllOraculPlaces };