+++ src/constants/Colors.jsx
@@ -4,7 +4,10 @@ export default {
   goldeenLight: "#ffc57b",
   goldeenMiddle: "#ee7b41",
   stone100: "#f5f5f4",
+  stone200: "#e7e5e4",
   stone300: "#d6d3d1",
   stone400: "#a8a29e",
+  stone700: "#44403c",
+  stone800: "#292524",
   red600: "#dc2626"
 };